Resolved merge conflicts:
- Keep ClassroomDetailForum naming (over Announcement)
- Keep forum-kelas route (over pengumuman)
- Integrate remote backend changes (pagination, policies, view count)
- Fix test files for new service constructors
- Clean up stale Announcement/AnnouncementDetail directories

+++ lms-usti-be/data/classroom.go
@@ -62,7 +62,7 @@ type AnnouncementResponse struct {
 	ClassroomName string `json:"classroom_name"`
 	CreatedBy     string `json:"created_by"`
 	CreatedAt     string `json:"created_at"`
-	CommentCount  int    `json:"comment_count"`
+	CommentCount  int64  `json:"comment_count"`
 }
 
 type ClassroomResponse struct {
